Update LinkedList Problems.py
The Edge cases got me on this problem. I forgot to add the case for if l1 and l2 had the same value, causing the code to run infinitely in that case, and then i neglected to use a dummy node to prevent the return value from being None. I should've been more detailed with my outline of potential cases to solve for.